@Iwasawafag 2016-12-04
You may notice that code performing rotation by 90 degrees moved to a separate function. I had to do that, otherwise performance degraded seriously in benchmark after rotation by 180 degrees. Something is wrong with V8's optimizator, it seems. It was becoming even slower than old implementation, that performed 2 extra rotations