@hipstersmoothie 2018-08-05
closing in favor of #512